requirements.txt
opencv-python~=4.9.0.80
mediapipe~=0.10.11
numpy~=1.26.4
nltk
pillow>=10.3.0
types-Pillow
torch~=2.2.1
pandas~=2.2.1
pandas-stubs
matplotlib~=3.8.3
torchvision~=0.17.1
keras
seaborn
scikit-learn
tensorflow
pytest
pytest-cov
ufmt
usort
mdformat
mypy
pylint
stubs
toml~=0.10.2
torch 
torchvision
timm
